Move ios API into IOSNotifications
